orch: Remove old code
Removed old action code, program and examples.
