+++ src/orchestration/BUILD
@@ -62,14 +62,3 @@ rust_binary(
         ":liborchestration",
     ],
 )
-
-rust_binary(
-    name = "basic_new",
-    srcs = [
-        "examples/basic_new.rs",
-    ] + glob(["examples/common/**/*.rs"]),
-    visibility = ["//visibility:public"],
-    deps = _COMMON_DEPS + [
-        ":liborchestration",
-    ],
-)
