+++ src/orchestration/examples/catch_error.rs
@@ -15,15 +15,11 @@ use async_runtime::{runtime::async_runtime::AsyncRuntimeBuilder, scheduler::exec
 use foundation::prelude::*;
 use logging_tracing::TracingLibraryBuilder;
 use orchestration::{
-    actions::internal::{
-        catch::ErrorFilter,
-        invoke::Invoke,
-        prelude::{CatchBuilder, ConcurrencyBuilder},
-        sequence::SequenceBuilder,
-    },
+    actions::{catch::ErrorFilter, invoke::Invoke, sequence::SequenceBuilder},
     api::{design::Design, Orchestration},
     common::DesignConfig,
-    program::internal::ProgramBuilder,
+    prelude::{CatchBuilder, ConcurrencyBuilder},
+    program::ProgramBuilder,
 };
 
 mod common;
