--- src/orchestration/src/actions/internal/mod.rs
@@ -1,22 +1,0 @@
-//
-// Copyright (c) 2025 Contributors to the Eclipse Foundation
-//
-// See the NOTICE file(s) distributed with this work for additional
-// information regarding copyright ownership.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Apache License Version 2.0 which is available at
-// <https://www.apache.org/licenses/LICENSE-2.0>
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-pub mod action;
-pub mod catch;
-pub mod concurrency;
-pub mod invoke;
-pub mod sequence;
-pub mod sync;
-pub mod trigger;
-
-pub mod prelude;
