+++ src/orchestration/src/actions/invoke.rs
@@ -11,123 +11,167 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-use super::action::{ActionResult, ActionTrait, ClonableActionTrait};
-use crate::core::runtime_seq_acc::RuntimeSequentialAccess;
-use async_runtime::core::types::box_future;
+use super::action::{ActionBaseMeta, ActionExecError, ActionResult, ActionTrait, ReusableBoxFutureResult, UserErrValue};
+use crate::{
+    api::design::Design,
+    common::{orch_tag::OrchestrationTag, tag::Tag},
+};
+use async_runtime::{
+    core::types::UniqueWorkerId, futures::reusable_box_future::ReusableBoxFuture, futures::reusable_box_future::ReusableBoxFuturePool,
+};
+use foundation::prelude::CommonErrors;
 use std::{
     future::Future,
     sync::{Arc, Mutex},
 };
 
-pub type FunctionType = fn() -> ActionResult;
+#[cfg(not(feature = "runtime-api-mock"))]
+use async_runtime::safety::spawn_from_reusable_on_dedicated;
+#[cfg(feature = "runtime-api-mock")]
+use async_runtime::testing::mock::spawn_from_reusable_on_dedicated;
 
-///
-/// TODO: Capturing object (owned, arced, closure) seems to be tricky from runtime perspective. So in theory, we will never call it in parallel, but if there is a timeout and user resumes program, we can come back to the same Invoke while other was not yet aborted (clogged, whatever).
-/// I have a feeling that we shall build handling mechanism for that in Orchestration. The same issue will happen with Computation and we need common strategy. Options:
-///  - panic on such case
-///  - trigger global error handler and stop program without resume possibility
-///  - block to wait until previous call did not finished - this may retrigger error handler again and again  ?
-///
-pub struct Invoke {} // Dummy struct to create fake namespace Invoke:: to hide types for impl details
+/// A result of an invoke action.
+pub type InvokeResult = Result<(), UserErrValue>;
+pub(crate) type InvokeFunctionType = fn() -> InvokeResult;
+
+pub struct Invoke {}
 
 impl Invoke {
-    ///
-    /// Creates Invoke action from a function pointer
-    ///
-    pub fn from_fn(func: FunctionType) -> Box<dyn ClonableActionTrait> {
-        Box::new(InvokeFn { action: func })
+    /// Create an invoke action out of an orchestration tag.
+    pub fn from_tag(tag: &OrchestrationTag) -> Box<dyn ActionTrait> {
+        (*tag.action_provider()).borrow_mut().provide_invoke(*tag.key()).unwrap()
     }
 
-    ///
-    /// Creates Invoke action from a non-clonable async function
-    ///
-    pub fn from_async<F, Fut>(function: F) -> Box<dyn ActionTrait>
+    pub fn from_design(name: &str, design: &Design) -> Box<dyn ActionTrait> {
+        let tag = design.get_orchestration_tag(name.into());
+        assert!(
+            tag.is_ok(),
+            "Failed to create invoke with name '{}', design/deployment errors where not handled properly before or You passing wrong name. ({:?})",
+            name,
+            tag
+        );
+
+        Self::from_tag(&tag.unwrap())
+    }
+
+    pub(crate) fn from_fn(tag: Tag, action: InvokeFunctionType, worker_id: Option<UniqueWorkerId>) -> Box<dyn ActionTrait> {
+        Box::new(InvokeFn {
+            action,
+            action_future_pool: ReusableBoxFuturePool::new(8, InvokeFn::action_future(action)),
+            worker_id,
+            base: ActionBaseMeta {
+                tag,
+                reusable_future_pool: ReusableBoxFuturePool::new(8, InvokeFn::spawn_action(InstantOrSpawn::None)),
+            },
+        })
+    }
+
+    pub(crate) fn from_async<A, F>(tag: Tag, action: A, worker_id: Option<UniqueWorkerId>) -> Box<dyn ActionTrait>
     where
-        F: FnMut() -> Fut + 'static + Send,
-        Fut: Future<Output = ActionResult> + 'static + Send,
+        A: Fn() -> F + 'static + Send,
+        F: Future<Output = InvokeResult> + 'static + Send,
     {
-        Box::new(InvokeAsyncFn { action: function })
+        let future = action();
+
+        Box::new(InvokeAsync {
+            action,
+            action_future_pool: ReusableBoxFuturePool::new(8, InvokeAsync::<A, F>::action_future(future)),
+            worker_id,
+            base: ActionBaseMeta {
+                tag,
+                reusable_future_pool: ReusableBoxFuturePool::new(8, InvokeAsync::<A, F>::spawn_action(InstantOrSpawn::None)),
+            },
+        })
     }
 
-    ///
-    /// Creates Invoke action from a clonable async function
-    ///
-    pub fn from_async_clonable<F, Fut>(function: F) -> Box<dyn ClonableActionTrait>
-    where
-        F: FnMut() -> Fut + 'static + Send + Clone,
-        Fut: Future<Output = ActionResult> + 'static + Send + Clone,
-    {
-        Box::new(InvokeAsyncFn { action: function })
-    }
-
-    pub fn from_not_owned<T: 'static>(
-        obj: Arc<RuntimeSequentialAccess<T>>,
-        method: fn(&mut Arc<RuntimeSequentialAccess<T>>) -> ActionResult,
+    pub(crate) fn from_method<T: 'static + Send>(
+        tag: Tag,
+        object: Arc<Mutex<T>>,
+        method: fn(&mut T) -> InvokeResult,
+        worker_id: Option<UniqueWorkerId>,
     ) -> Box<dyn ActionTrait> {
-        Box::new(InvokeOwnedObject {
-            object: Arc::new(RuntimeSequentialAccess::new(obj)),
-            method: FnType::Mut(method),
+        Box::new(InvokeMethod {
+            object: Arc::clone(&object),
+            method,
+            action_future_pool: ReusableBoxFuturePool::new(8, InvokeMethod::<T>::action_future(Arc::clone(&object), method)),
+            worker_id,
+            base: ActionBaseMeta {
+                tag,
+                reusable_future_pool: ReusableBoxFuturePool::new(8, InvokeMethod::<T>::spawn_action(InstantOrSpawn::None)),
+            },
         })
     }
 
-    pub fn from_arc<T: 'static + Send>(obj: Arc<Mutex<T>>, method: fn(&mut T) -> ActionResult) -> Box<dyn ClonableActionTrait> {
-        Box::new(InvokeArc {
-            object: obj,
-            method: FnType::Mut(method),
-        })
-    }
-
-    pub fn from_arc_mtx<T: 'static + Send, F, Fut>(obj: Arc<Mutex<T>>, method: F) -> Box<dyn ActionTrait>
+    pub(crate) fn from_method_async<T, M, F>(tag: Tag, object: Arc<Mutex<T>>, method: M, worker_id: Option<UniqueWorkerId>) -> Box<dyn ActionTrait>
     where
-        F: FnMut(Arc<Mutex<T>>) -> Fut + 'static + Send,
-        Fut: Future<Output = ActionResult> + 'static + Send,
+        T: 'static + Send,
+        M: Fn(Arc<Mutex<T>>) -> F + 'static + Send,
+        F: Future<Output = InvokeResult> + 'static + Send,
     {
-        Box::new(InvokeArcMtx { object: obj, method })
-    }
+        let future = (method)(Arc::clone(&object));
 
-    pub fn from_arc_mtx_clonable<T: 'static + Send, F, Fut>(obj: Arc<Mutex<T>>, method: F) -> Box<dyn ClonableActionTrait>
-    where
-        F: FnMut(Arc<Mutex<T>>) -> Fut + 'static + Send + Clone,
-        Fut: Future<Output = ActionResult> + 'static + Send + Clone,
-    {
-        Box::new(InvokeArcMtx { object: obj, method })
-    }
-
-    /// # Safety
-    ///
-    /// TODO: This is marked as unsafe right now (as precaution indicator). User need to know restrictions before doing this. Check [`InvokeOwnedObject`].
-    pub unsafe fn from_owned<T: 'static>(obj: T, method: fn(&T) -> ActionResult) -> Box<dyn ActionTrait> {
-        Box::new(InvokeOwnedObject {
-            object: Arc::new(RuntimeSequentialAccess::new(obj)),
-            method: FnType::Immut(method),
-        })
-    }
-
-    /// # Safety
-    ///
-    /// TODO: This is marked as unsafe right now (as precaution indicator). User need to know restrictions before doing this. Check [`InvokeOwnedObject`].
-    pub unsafe fn from_owned_mut<T: 'static>(obj: T, method: fn(&mut T) -> ActionResult) -> Box<dyn ActionTrait> {
-        Box::new(InvokeOwnedObject {
-            object: Arc::new(RuntimeSequentialAccess::new(obj)),
-            method: FnType::Mut(method),
+        Box::new(InvokeMethodAsync {
+            object,
+            method,
+            action_future_pool: ReusableBoxFuturePool::new(8, InvokeMethodAsync::<T, M, F>::action_future(future)),
+            worker_id,
+            base: ActionBaseMeta {
+                tag,
+                reusable_future_pool: ReusableBoxFuturePool::new(8, InvokeMethodAsync::<T, M, F>::spawn_action(InstantOrSpawn::None)),
+            },
         })
     }
 }
 
-#[derive(Clone)]
+fn invoke_result_into_action_result(result: InvokeResult) -> ActionResult {
+    result.map_err(|err| err.into())
+}
+
+enum InstantOrSpawn<I> {
+    None,
+    Instant(I),
+    Spawn(ReusableBoxFuture<ActionResult>, UniqueWorkerId),
+}
+
 struct InvokeFn {
-    action: FunctionType,
+    action: InvokeFunctionType,
+    action_future_pool: ReusableBoxFuturePool<ActionResult>,
+    worker_id: Option<UniqueWorkerId>,
+    base: ActionBaseMeta,
 }
 
 impl InvokeFn {
-    async fn internal_future(action: FunctionType) -> ActionResult {
-        action()
+    async fn action_future(action: InvokeFunctionType) -> ActionResult {
+        invoke_result_into_action_result(action())
+    }
+
+    async fn spawn_action(instant_or_spawn: InstantOrSpawn<InvokeFunctionType>) -> ActionResult {
+        match instant_or_spawn {
+            InstantOrSpawn::None => Ok(()),
+            InstantOrSpawn::Instant(action) => invoke_result_into_action_result(action()),
+            InstantOrSpawn::Spawn(future, worker_id) => match spawn_from_reusable_on_dedicated(future, worker_id).await {
+                Ok(result) => result,
+                Err(_) => Err(ActionExecError::Internal),
+            },
+        }
     }
 }
 
 impl ActionTrait for InvokeFn {
-    fn execute(&mut self) -> super::action::ActionFuture {
-        box_future(Self::internal_future(self.action))
+    fn try_execute(&mut self) -> ReusableBoxFutureResult {
+        if let Some(worker_id) = self.worker_id {
+            match self.action_future_pool.next(InvokeFn::action_future(self.action)) {
+                Ok(future) => self
+                    .base
+                    .reusable_future_pool
+                    .next(InvokeFn::spawn_action(InstantOrSpawn::Spawn(future, worker_id))),
+                Err(_) => Err(CommonErrors::GenericError),
+            }
+        } else {
+            self.base
+                .reusable_future_pool
+                .next(InvokeFn::spawn_action(InstantOrSpawn::Instant(self.action)))
+        }
     }
 
     fn name(&self) -> &'static str {
@@ -135,239 +179,283 @@ impl ActionTrait for InvokeFn {
     }
 
     fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let indent = " ".repeat(nest);
-        writeln!(f, "{}|-{}", indent, self.name())
-    }
-
-    fn fill_runtime_info(&mut self, _p: &mut super::action::ActionRuntimeInfoProvider) {}
-}
-
-impl ClonableActionTrait for InvokeFn {
-    fn clone_boxed<'a>(&self) -> Box<dyn ClonableActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        Box::new(self.clone())
-    }
-
-    fn into_boxed_action<'a>(self: Box<Self>) -> Box<dyn ActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        self
+        writeln!(f, "{}|-{}", " ".repeat(nest), self.name())
     }
 }
 
-#[derive(Clone)]
-struct InvokeAsyncFn<T, Fut>
+struct InvokeAsync<A, F>
 where
-    T: FnMut() -> Fut + 'static + Send,
-    Fut: Future<Output = ActionResult> + 'static + Send,
+    A: Fn() -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
 {
-    action: T,
+    action: A,
+    action_future_pool: ReusableBoxFuturePool<ActionResult>,
+    worker_id: Option<UniqueWorkerId>,
+    base: ActionBaseMeta,
 }
 
-impl<T, Fut> ActionTrait for InvokeAsyncFn<T, Fut>
+impl<A, F> InvokeAsync<A, F>
 where
-    T: FnMut() -> Fut + 'static + Send,
-    Fut: Future<Output = ActionResult> + 'static + Send,
+    A: Fn() -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
 {
-    fn execute(&mut self) -> super::action::ActionFuture {
-        let fut = (self.action)();
-        box_future(fut)
+    async fn action_future(future: F) -> ActionResult {
+        invoke_result_into_action_result(future.await)
+    }
+
+    async fn spawn_action(instant_or_spawn: InstantOrSpawn<F>) -> ActionResult {
+        match instant_or_spawn {
+            InstantOrSpawn::None => Ok(()),
+            InstantOrSpawn::Instant(action) => invoke_result_into_action_result(action.await),
+            InstantOrSpawn::Spawn(future, worker_id) => match spawn_from_reusable_on_dedicated(future, worker_id).await {
+                Ok(result) => result,
+                Err(_) => Err(ActionExecError::Internal),
+            },
+        }
+    }
+}
+
+impl<A, F> ActionTrait for InvokeAsync<A, F>
+where
+    A: Fn() -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
+{
+    fn try_execute(&mut self) -> ReusableBoxFutureResult {
+        if let Some(worker_id) = self.worker_id {
+            match self.action_future_pool.next(InvokeAsync::<A, F>::action_future((self.action)())) {
+                Ok(future) => self
+                    .base
+                    .reusable_future_pool
+                    .next(InvokeAsync::<A, F>::spawn_action(InstantOrSpawn::Spawn(future, worker_id))),
+                Err(_) => Err(CommonErrors::GenericError),
+            }
+        } else {
+            self.base
+                .reusable_future_pool
+                .next(InvokeAsync::<A, F>::spawn_action(InstantOrSpawn::Instant((self.action)())))
+        }
     }
 
     fn name(&self) -> &'static str {
-        "Invoke"
+        "InvokeAsync"
     }
 
     fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let indent = " ".repeat(nest);
-        writeln!(f, "{}|-{}", indent, self.name())
-    }
-
-    fn fill_runtime_info(&mut self, _p: &mut super::action::ActionRuntimeInfoProvider) {}
-}
-
-impl<T, Fut> ClonableActionTrait for InvokeAsyncFn<T, Fut>
-where
-    T: FnMut() -> Fut + 'static + Send + Clone,
-    Fut: Future<Output = ActionResult> + 'static + Send + Clone,
-{
-    fn clone_boxed<'a>(&self) -> Box<dyn ClonableActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        Box::new(self.clone())
-    }
-
-    fn into_boxed_action<'a>(self: Box<Self>) -> Box<dyn ActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        self
+        writeln!(f, "{}|-{}", " ".repeat(nest), self.name())
     }
 }
 
-enum FnType<T> {
-    Mut(fn(&mut T) -> ActionResult),
-    Immut(fn(&T) -> ActionResult),
-}
+type InvokeMethodType<T> = fn(&mut T) -> InvokeResult;
 
-impl<T> Clone for FnType<T> {
-    fn clone(&self) -> Self {
-        match self {
-            Self::Mut(arg0) => Self::Mut(*arg0),
-            Self::Immut(arg0) => Self::Immut(*arg0),
-        }
-    }
-}
-
-///
-/// TODO: Decide if we are going for it....
-/// If we do own object, the Orchestration assures that it will never be called in parallel, so we can take over immutable leash from Arc.
-/// Safety:
-///     - The T implementer must ensure that provides method will not block. Otherwise next attempt for this execution will panic!
-///
-struct InvokeOwnedObject<T: 'static> {
-    object: Arc<RuntimeSequentialAccess<T>>,
-    method: FnType<T>,
-}
-
-impl<T> InvokeOwnedObject<T> {
-    async fn internal_future(object: Arc<RuntimeSequentialAccess<T>>, method: FnType<T>) -> ActionResult {
-        let mut guard = object.lock();
-
-        match method {
-            FnType::Mut(callable) => (callable)(&mut guard),
-            FnType::Immut(callable) => (callable)(&guard),
-        }
-    }
-}
-
-impl<T> ActionTrait for InvokeOwnedObject<T> {
-    fn execute(&mut self) -> super::action::ActionFuture {
-        // TODO: We need to panic here if we have it timeout on this method and program returned back here. This is bad implementation on user side.
-        if self.object.is_locked() {
-            panic!("This action is still running, we cannot let it run again. This clearly indicate bug int user application layer")
-        }
-
-        box_future(Self::internal_future(self.object.clone(), self.method.clone()))
-    }
-
-    fn name(&self) -> &'static str {
-        "Invoke"
-    }
-
-    fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let indent = " ".repeat(nest);
-        writeln!(f, "{}|-{}", indent, self.name())
-    }
-
-    fn fill_runtime_info(&mut self, _p: &mut super::action::ActionRuntimeInfoProvider) {}
-}
-
-struct InvokeArc<T: 'static> {
+struct InvokeMethod<T: 'static + Send> {
     object: Arc<Mutex<T>>,
-    method: FnType<T>,
+    method: InvokeMethodType<T>,
+    action_future_pool: ReusableBoxFuturePool<ActionResult>,
+    worker_id: Option<UniqueWorkerId>,
+    base: ActionBaseMeta,
 }
 
-impl<T: 'static> InvokeArc<T> {
-    async fn internal_future(object: Arc<Mutex<T>>, method: FnType<T>) -> ActionResult {
-        let mut guard = object.lock().unwrap();
+impl<T: 'static + Send> InvokeMethod<T> {
+    async fn action_future(object: Arc<Mutex<T>>, method: InvokeMethodType<T>) -> ActionResult {
+        let mut object = object.lock().unwrap();
+        invoke_result_into_action_result(method(&mut object))
+    }
 
-        match method {
-            FnType::Mut(callable) => (callable)(&mut guard),
-            FnType::Immut(callable) => (callable)(&guard),
+    async fn spawn_action(instant_or_spawn: InstantOrSpawn<(Arc<Mutex<T>>, InvokeMethodType<T>)>) -> ActionResult {
+        match instant_or_spawn {
+            InstantOrSpawn::None => Ok(()),
+            InstantOrSpawn::Instant((object, method)) => {
+                let mut object = object.lock().unwrap();
+                invoke_result_into_action_result(method(&mut object))
+            }
+            InstantOrSpawn::Spawn(future, worker_id) => match spawn_from_reusable_on_dedicated(future, worker_id).await {
+                Ok(result) => result,
+                Err(_) => Err(ActionExecError::Internal),
+            },
         }
     }
 }
 
-impl<T: 'static + Send> ActionTrait for InvokeArc<T> {
-    fn execute(&mut self) -> super::action::ActionFuture {
-        box_future(Self::internal_future(self.object.clone(), self.method.clone()))
+impl<T: 'static + Send> ActionTrait for InvokeMethod<T> {
+    fn try_execute(&mut self) -> ReusableBoxFutureResult {
+        if let Some(worker_id) = self.worker_id {
+            match self
+                .action_future_pool
+                .next(InvokeMethod::<T>::action_future(Arc::clone(&self.object), self.method))
+            {
+                Ok(future) => self
+                    .base
+                    .reusable_future_pool
+                    .next(InvokeMethod::<T>::spawn_action(InstantOrSpawn::Spawn(future, worker_id))),
+                Err(_) => Err(CommonErrors::GenericError),
+            }
+        } else {
+            self.base
+                .reusable_future_pool
+                .next(InvokeMethod::<T>::spawn_action(InstantOrSpawn::Instant((
+                    Arc::clone(&self.object),
+                    self.method,
+                ))))
+        }
     }
 
     fn name(&self) -> &'static str {
-        "Invoke"
+        "InvokeAsync"
     }
 
     fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let indent = " ".repeat(nest);
-        writeln!(f, "{}|-{}", indent, self.name())
-    }
-
-    fn fill_runtime_info(&mut self, _p: &mut super::action::ActionRuntimeInfoProvider) {}
-}
-
-impl<T: 'static + Send> ClonableActionTrait for InvokeArc<T> {
-    fn clone_boxed<'a>(&self) -> Box<dyn ClonableActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        Box::new(InvokeArc {
-            object: Arc::clone(&self.object),
-            method: self.method.clone(),
-        })
-    }
-
-    fn into_boxed_action<'a>(self: Box<Self>) -> Box<dyn ActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        self
+        writeln!(f, "{}|-{}", " ".repeat(nest), self.name())
     }
 }
 
-struct InvokeArcMtx<T: 'static, F, Fut>
+struct InvokeMethodAsync<T, M, F>
 where
-    F: FnMut(Arc<Mutex<T>>) -> Fut + 'static + Send,
-    Fut: Future<Output = ActionResult> + 'static + Send,
+    T: 'static + Send,
+    M: FnMut(Arc<Mutex<T>>) -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
 {
     object: Arc<Mutex<T>>,
-    method: F,
+    method: M,
+    action_future_pool: ReusableBoxFuturePool<ActionResult>,
+    worker_id: Option<UniqueWorkerId>,
+    base: ActionBaseMeta,
 }
 
-impl<T: 'static + Send, F, Fut> ActionTrait for InvokeArcMtx<T, F, Fut>
+impl<T, M, F> InvokeMethodAsync<T, M, F>
 where
-    F: FnMut(Arc<Mutex<T>>) -> Fut + 'static + Send,
-    Fut: Future<Output = ActionResult> + 'static + Send,
+    T: 'static + Send,
+    M: FnMut(Arc<Mutex<T>>) -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
 {
-    fn execute(&mut self) -> super::action::ActionFuture {
-        box_future((self.method)(self.object.clone()))
+    async fn action_future(future: F) -> ActionResult {
+        invoke_result_into_action_result(future.await)
     }
 
+    async fn spawn_action(instant_or_spawn: InstantOrSpawn<F>) -> ActionResult {
+        match instant_or_spawn {
+            InstantOrSpawn::None => Ok(()),
+            InstantOrSpawn::Instant(future) => invoke_result_into_action_result(future.await),
+            InstantOrSpawn::Spawn(future, worker_id) => match spawn_from_reusable_on_dedicated(future, worker_id).await {
+                Ok(result) => result,
+                Err(_) => Err(ActionExecError::Internal),
+            },
+        }
+    }
+}
+
+impl<T, M, F> ActionTrait for InvokeMethodAsync<T, M, F>
+where
+    T: 'static + Send,
+    M: FnMut(Arc<Mutex<T>>) -> F + 'static + Send,
+    F: Future<Output = InvokeResult> + 'static + Send,
+{
+    fn try_execute(&mut self) -> ReusableBoxFutureResult {
+        if let Some(worker_id) = self.worker_id {
+            match self
+                .action_future_pool
+                .next(InvokeMethodAsync::<T, M, F>::action_future((self.method)(Arc::clone(&self.object))))
+            {
+                Ok(future) => self
+                    .base
+                    .reusable_future_pool
+                    .next(InvokeMethodAsync::<T, M, F>::spawn_action(InstantOrSpawn::Spawn(future, worker_id))),
+                Err(_) => Err(CommonErrors::GenericError),
+            }
+        } else {
+            self.base
+                .reusable_future_pool
+                .next(InvokeMethodAsync::<T, M, F>::spawn_action(InstantOrSpawn::Instant((self.method)(
+                    Arc::clone(&self.object),
+                ))))
+        }
+    }
     fn name(&self) -> &'static str {
-        "Invoke"
+        "InvokeAsync"
     }
 
     fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let indent = " ".repeat(nest);
-        writeln!(f, "{}|-{}", indent, self.name())
+        writeln!(f, "{}|-{}", " ".repeat(nest), self.name())
     }
-
-    fn fill_runtime_info(&mut self, _p: &mut super::action::ActionRuntimeInfoProvider) {}
 }
 
-impl<T: 'static + Send, F, Fut> ClonableActionTrait for InvokeArcMtx<T, F, Fut>
-where
-    F: FnMut(Arc<Mutex<T>>) -> Fut + 'static + Send + Clone,
-    Fut: Future<Output = ActionResult> + 'static + Send + Clone,
-{
-    fn clone_boxed<'a>(&self) -> Box<dyn ClonableActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        Box::new(InvokeArcMtx {
-            object: Arc::clone(&self.object),
-            method: self.method.clone(),
-        })
+#[cfg(test)]
+#[cfg(not(loom))]
+mod tests {
+    use std::sync::{Arc, Mutex};
+
+    #[test]
+    fn test_fn() {
+        fn test() -> super::InvokeResult {
+            Ok(())
+        }
+
+        // Capture the same action multiple times.
+        let mut action1 = super::Invoke::from_fn("tag".into(), test, None);
+        let mut action2 = super::Invoke::from_fn("tag".into(), test, None);
+        // Execute the same invoke multiple times.
+        assert!(action1.try_execute().is_ok());
+        assert!(action1.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
     }
 
-    fn into_boxed_action<'a>(self: Box<Self>) -> Box<dyn ActionTrait + 'a>
-    where
-        Self: 'a,
-    {
-        self
+    #[test]
+    fn test_async() {
+        async fn test() -> super::InvokeResult {
+            Ok(())
+        }
+
+        // Capture the same action multiple times.
+        let mut action1 = super::Invoke::from_async("tag".into(), test, None);
+        let mut action2 = super::Invoke::from_async("tag".into(), test, None);
+        // Execute the same invoke multiple times.
+        assert!(action1.try_execute().is_ok());
+        assert!(action1.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+    }
+
+    #[test]
+    fn test_method() {
+        struct TestObject {}
+
+        impl TestObject {
+            fn test_method(&mut self) -> super::InvokeResult {
+                Err(0xcafe_u64.into())
+            }
+        }
+
+        let object = Arc::new(Mutex::new(TestObject {}));
+
+        // Capture the same action multiple times.
+        let mut action1 = super::Invoke::from_method("tag".into(), Arc::clone(&object), TestObject::test_method, None);
+        let mut action2 = super::Invoke::from_method("tag".into(), Arc::clone(&object), TestObject::test_method, None);
+        // Execute the same invoke multiple times.
+        assert!(action1.try_execute().is_ok());
+        assert!(action1.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+    }
+
+    #[test]
+    fn test_method_async() {
+        struct TestObject {}
+
+        async fn test_method(object: Arc<Mutex<TestObject>>) -> super::InvokeResult {
+            let _guard = object.lock().unwrap();
+
+            Err(0xcafe_u64.into())
+        }
+
+        let object = Arc::new(Mutex::new(TestObject {}));
+
+        // Capture the same action multiple times.
+        let mut action1 = super::Invoke::from_method_async("tag".into(), Arc::clone(&object), test_method, None);
+        let mut action2 = super::Invoke::from_method_async("tag".into(), Arc::clone(&object), test_method, None);
+        // Execute the same invoke multiple times.
+        assert!(action1.try_execute().is_ok());
+        assert!(action1.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
+        assert!(action2.try_execute().is_ok());
     }
 }
