+++ src/orchestration/src/actions/mod.rs
@@ -12,9 +12,8 @@
 //
 
 pub mod action;
+pub mod catch;
 pub mod concurrency;
-pub mod event;
-pub mod internal;
 pub mod invoke;
 pub mod sequence;
 pub mod sync;
