+++ src/orchestration/src/actions/sequence.rs
@@ -11,70 +11,157 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-use async_runtime::core::types::box_future;
-use logging_tracing::prelude::*;
+use super::action::{ActionBaseMeta, ActionResult, ActionTrait, ReusableBoxFutureResult};
+use crate::common::tag::Tag;
 
-use super::action::{ActionBaseMeta, ActionFuture, ActionResult, ActionTrait, NamedId};
+use async_runtime::futures::reusable_box_future::{ReusableBoxFuture, ReusableBoxFuturePool};
+use foundation::{
+    containers::{growable_vec::GrowableVec, reusable_objects::ReusableObject, reusable_vec_pool::ReusableVecPool},
+    prelude::*,
+};
 
-pub struct Sequence {
-    base: ActionBaseMeta,
-    actions: Vec<Box<dyn ActionTrait>>,
+const REUSABLE_FUTURE_POOL_SIZE: usize = 2;
+const REUSABLE_VEC_POOL_SIZE: usize = 2;
+const DEFAULT_TAG: &str = "orch::internal::sequence";
+
+///
+/// Construct a `SequenceBuilder` for creating a `Sequence` action
+///
+pub struct SequenceBuilder {
+    actions: GrowableVec<Box<dyn ActionTrait>>,
 }
 
-impl Sequence {
-    /// Create a new sequence action
-    pub fn new() -> Box<Sequence> {
-        Self::new_internal(NamedId::default())
+impl Default for SequenceBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SequenceBuilder {
+    ///
+    /// Construct a `SequenceBuilder`
+    ///
+    pub fn new() -> SequenceBuilder {
+        const REUSABLE_VEC_SIZE: usize = 4;
+        Self {
+            actions: GrowableVec::new(REUSABLE_VEC_SIZE),
+        }
     }
 
-    pub fn new_with_id(id: NamedId) -> Box<Sequence> {
-        Self::new_internal(id)
-    }
-
-    fn new_internal(named_id: NamedId) -> Box<Sequence> {
-        Box::new(Self {
-            actions: Vec::new(),
-            base: ActionBaseMeta {
-                named_id,
-                runtime: Default::default(),
-            },
-        })
-    }
-
-    /// Add sequence step
-    pub fn with_step(mut self: Box<Self>, action: Box<dyn ActionTrait>) -> Box<Self> {
+    ///
+    /// Add an action to the `Sequence`
+    ///
+    pub fn with_step(&mut self, action: Box<dyn ActionTrait>) -> &mut Self {
         self.actions.push(action);
         self
     }
 
-    /// Execute a futures collection and terminates immediately upon error
-    async fn execute_impl(meta: ActionBaseMeta, mut futures: Vec<ActionFuture>) -> ActionResult {
-        let mut result = Ok(());
+    ///
+    /// Build the `Sequence` action
+    ///
+    /// # Errors
+    ///
+    /// Returns ``Err(NoData)` if no reusable futures collection for storing the actions' futures is available
+    ///
+    /// # Panics
+    ///
+    /// Panics if the `Sequence` does not contain any actions
+    ///
+    pub fn build(&mut self) -> Box<Sequence> {
+        assert!(!self.actions.is_empty(), "Sequence must contain at least one action!");
 
-        trace!(sequence = ?meta, "Before joining steps");
+        // No more actions may be added beyond this point
+        self.actions.lock();
 
-        for future in futures.iter_mut() {
-            result = future.await;
-            if result.is_err() {
-                // terminate sequence and propagate the error
-                error!("error in step!");
-                break;
-            }
+        // Create pools
+        let (futures_vec_pool, reusable_future_pool) = SequenceBuilder::create_pools(self.actions.len());
+
+        // Move the actions from Builder's GrowableVec to Sequence's fixed-sized Vec
+        // Here we also reverse the order, so that the actions become already in the correct order,
+        // when they are popped out in the execute_impl() later on
+        let mut actions = Vec::<Box<dyn ActionTrait>>::new(self.actions.len());
+        while let Some(action) = self.actions.pop() {
+            actions.push(action);
         }
 
-        trace!(sequence = ?meta, "After joining steps");
-        result
+        // Finally, return the `Sequence` action
+        Box::new(Sequence {
+            actions,
+            base: ActionBaseMeta {
+                tag: Tag::from_str_static(DEFAULT_TAG),
+                reusable_future_pool,
+            },
+            futures_vec_pool,
+        })
+    }
+
+    ///
+    /// Create pools of reusable futures vec and reusable future
+    ///
+    fn create_pools(futures_size: usize) -> (ReusableVecPool<ReusableBoxFuture<ActionResult>>, ReusableBoxFuturePool<ActionResult>) {
+        let mut futures_vec_pool = ReusableVecPool::<ReusableBoxFuture<ActionResult>>::new(REUSABLE_VEC_POOL_SIZE, |_| Vec::new(futures_size));
+        let futures_vec = futures_vec_pool.next_object().unwrap();
+
+        // Populate the futures' collection to initialize the reusable future pool's layout
+        let reusable_future_pool = ReusableBoxFuturePool::<ActionResult>::new(
+            REUSABLE_FUTURE_POOL_SIZE,
+            Sequence::execute_impl(Tag::from_str_static(DEFAULT_TAG), futures_vec),
+        );
+
+        (futures_vec_pool, reusable_future_pool)
+    }
+}
+
+///
+/// An orchestration action that invokes subsequent actions specified via `with_step()` in a FIFO
+/// manner.
+///
+/// If any action encounters an error, the `Sequence` execution will terminate immediately,
+/// preventing the execution of any remaining actions.
+///
+pub struct Sequence {
+    actions: Vec<Box<dyn ActionTrait>>,
+    base: ActionBaseMeta,
+    futures_vec_pool: ReusableVecPool<ReusableBoxFuture<ActionResult>>,
+}
+
+impl Sequence {
+    async fn execute_impl(tag: Tag, mut futures: ReusableObject<Vec<ReusableBoxFuture<ActionResult>>>) -> ActionResult {
+        // Execute all futures in the collection, but terminates immediately upon error
+        // We can directly pop() without reversing the order here, because the reversion already took place
+        // during elements transfer from Builder's GrowableVec to Sequence's Vec
+        while let Some(future) = futures.pop() {
+            trace!(step = ?tag, "Before awaiting step");
+            let result = future.into_pin().await;
+            if result.is_err() {
+                // Terminate sequence and propagate the error
+                error!("Error in sequence step {:?}", tag);
+                return result;
+            }
+            trace!(step = ?tag, "After awaiting step");
+        }
+
+        Ok(())
     }
 }
 
 impl ActionTrait for Sequence {
-    /// Will be called on each sequence step
-    fn execute(&mut self) -> ActionFuture {
-        // first collect all futures from the steps
-        let futures: std::vec::Vec<_> = self.actions.iter_mut().map(|action| action.execute()).collect();
+    fn try_execute(&mut self) -> ReusableBoxFutureResult {
+        // Get a fresh reusable futures collection and re-populate it with actions' futures
+        let mut futures_vec_pool = self.futures_vec_pool.next_object()?;
+        self.actions.iter_mut().try_for_each(|action| {
+            // Return error in (unlikely) case that no more future can be added to the reusable collection
+            if !futures_vec_pool.push(action.try_execute()?) {
+                error!("Unable to add a future to the reusable future vec in {:?}", self.base);
+                return Err(CommonErrors::NoSpaceLeft);
+            }
+            Ok(())
+        })?;
 
-        // and execute them
-        box_future(Sequence::execute_impl(self.base, futures))
+        // Get a future from the reusable future pool and execute it
+        self.base
+            .reusable_future_pool
+            .next(Sequence::execute_impl(self.base.tag, futures_vec_pool))
     }
 
     fn name(&self) -> &'static str {
@@ -84,87 +171,134 @@ impl ActionTrait for Sequence {
     fn dbg_fmt(&self, nest: usize, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let indent = " ".repeat(nest);
         writeln!(f, "{}|-{} - {:?}", indent, self.name(), self.base)?;
-        self.actions.iter().try_for_each(|x| {
+        self.actions.iter().try_for_each(|action| {
             writeln!(f, "{} |step", indent)?;
-            x.dbg_fmt(nest + 1, f)
+            action.dbg_fmt(nest + 1, f)
         })
-    }
-
-    fn fill_runtime_info(&mut self, p: &mut super::action::ActionRuntimeInfoProvider) {
-        self.base.runtime = p.next();
-        self.actions.iter_mut().for_each(|item| item.fill_runtime_info(p));
     }
 }
 
 #[cfg(test)]
+#[cfg(not(loom))]
 mod tests {
     use super::*;
-    use crate::actions::invoke::Invoke;
+    use crate::actions::action::{ActionExecError, UserErrValue};
+    use crate::testing::{MockActionBuilder, OrchTestingPoller};
+
     use std::task::Poll;
-    use testing::prelude::*;
-
-    async fn add_async<const N: usize>() -> ActionResult {
-        Ok(())
-    }
-
-    fn add_sync<const N: usize>() -> ActionResult {
-        Ok(())
-    }
 
     #[test]
-    fn test_with_future() {
-        let mut seq = Sequence::new()
-            .with_step(Invoke::from_async(add_async::<42>))
-            .with_step(Invoke::from_fn(add_sync::<2>).into_boxed_action());
+    #[should_panic]
+    fn build_on_empty_sequence_should_panic() {
+        // Create an empty sequence
+        let mut seq = SequenceBuilder::new().build();
 
-        // create mock with the resulting future
-        let mut mock = TestingFuturePoller::new(seq.execute());
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
         assert_eq!(Poll::Ready(Ok(())), mock.poll());
     }
 
     #[test]
-    fn test_with_boxed_future() {
-        let mut seq = Sequence::new()
-            .with_step(Invoke::from_async(add_async::<1729>))
-            .with_step(Invoke::from_fn(add_sync::<711>).into_boxed_action());
+    fn all_subsequent_steps_are_called() {
+        // Create a sequence with no nested branch
+        let mock_1 = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_2 = Box::new(MockActionBuilder::new().times(1).build());
+        let mut seq = SequenceBuilder::new().with_step(mock_1).with_step(mock_2).build();
 
-        // create mock with the resulting future as boxed future
-        let mut mock = TestingFuturePoller::new(Box::pin(seq.execute()));
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
         assert_eq!(Poll::Ready(Ok(())), mock.poll());
     }
 
     #[test]
-    fn test_nested() {
-        let mut seq = Sequence::new()
-            .with_step(Invoke::from_async(add_async::<42>))
+    fn all_steps_within_nested_steps_seq_are_called() {
+        // Create a sequence with a nested branch
+        let mock_1 = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_nested_a = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_nested_b = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_2 = Box::new(MockActionBuilder::new().times(1).build());
+        let mut seq = SequenceBuilder::new()
+            .with_step(mock_1)
+            .with_step(SequenceBuilder::new().with_step(mock_nested_a).with_step(mock_nested_b).build())
+            .with_step(mock_2)
+            .build();
+
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
+        assert_eq!(Poll::Ready(Ok(())), mock.poll());
+    }
+
+    #[test]
+    fn all_steps_within_double_nested_steps_seq_are_called() {
+        // Create a sequence with double nested branches
+        let mock_1 = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_2 = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_nested_2a = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_double_nested_2aa = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_double_nested_2ab = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_nested_2b = Box::new(MockActionBuilder::new().times(1).build());
+        let mut seq = SequenceBuilder::new()
+            .with_step(mock_1)
+            .with_step(mock_2)
             .with_step(
-                Sequence::new()
-                    .with_step(Invoke::from_async(add_async::<1729>))
+                SequenceBuilder::new()
+                    .with_step(mock_nested_2a)
                     .with_step(
-                        Sequence::new()
-                            .with_step(Invoke::from_fn(add_sync::<4242>).into_boxed_action())
-                            .with_step(Invoke::from_async(add_async::<2424>)),
+                        SequenceBuilder::new()
+                            .with_step(mock_double_nested_2aa)
+                            .with_step(mock_double_nested_2ab)
+                            .build(),
                     )
-                    .with_step(Invoke::from_fn(add_sync::<1881>).into_boxed_action()),
+                    .with_step(mock_nested_2b)
+                    .build(),
             )
-            .with_step(Invoke::from_async(add_async::<2>))
-            .with_step(Invoke::from_fn(add_sync::<711>).into_boxed_action());
+            .build();
 
-        // execute the action and check the results
-        let mut mock = TestingFuturePoller::new(seq.execute());
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
         assert_eq!(Poll::Ready(Ok(())), mock.poll());
     }
 
     #[test]
-    fn test_with_trackable_waker() {
-        let mut seq = Sequence::new().with_step(Invoke::from_fn(add_sync::<42>).into_boxed_action());
+    fn step_with_err_terminates_immediately() {
+        // Create a sequence that contains an error
+        let mock_ok = Box::new(MockActionBuilder::new().will_once(Ok(())).build());
+        let user_err = ActionExecError::UserError(UserErrValue::from(42));
+        let mock_err_1 = Box::new(MockActionBuilder::new().will_once(Err(user_err)).build());
+        let mock_err_2 = Box::new(MockActionBuilder::new().times(0).build());
+        let mut seq = SequenceBuilder::new()
+            .with_step(mock_ok)
+            .with_step(mock_err_1)
+            .with_step(mock_err_2)
+            .build();
 
-        // use trackable waker for polling
-        let waker = TrackableWaker::new();
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
+        assert_eq!(Poll::Ready(Err(user_err)), mock.poll());
+    }
 
-        // execute the action and check the results
-        let mut mock = TestingFuturePoller::new(seq.execute());
-        assert_eq!(Poll::Ready(Ok(())), mock.poll_with_waker(&waker.get_waker()));
-        assert_eq!(1, waker.get_waker_ref_count());
+    #[test]
+    fn step_with_err_in_nested_branch_terminates_immediately() {
+        // Create a sequence that contains an error step within a nested branch
+        let mock_1 = Box::new(MockActionBuilder::new().times(1).build());
+        let mock_2 = Box::new(MockActionBuilder::new().will_once(Ok(())).build());
+        let mock_nested_2a = Box::new(MockActionBuilder::new().will_once(Ok(())).build());
+        let mock_nested_err = Box::new(MockActionBuilder::new().will_once(Err(ActionExecError::NonRecoverableFailure)).build());
+        let mock_nested_2b = Box::new(MockActionBuilder::new().times(0).build());
+        let mut seq = SequenceBuilder::new()
+            .with_step(mock_1)
+            .with_step(mock_2)
+            .with_step(
+                SequenceBuilder::new()
+                    .with_step(mock_nested_2a)
+                    .with_step(mock_nested_err)
+                    .with_step(mock_nested_2b)
+                    .build(),
+            )
+            .build();
+
+        // Execute the sequence
+        let mut mock = OrchTestingPoller::new(seq.try_execute().unwrap());
+        assert_eq!(Poll::Ready(Err(ActionExecError::NonRecoverableFailure)), mock.poll());
     }
 }
