+++ src/orchestration/src/events/local_events.rs
@@ -18,7 +18,7 @@ use async_runtime::channels::spmc_broadcast::*;
 
 use crate::events::event_traits::{ListenerTrait, NotifierTrait};
 use crate::{
-    actions::internal::action::{ActionExecError, ActionResult},
+    actions::action::{ActionExecError, ActionResult},
     core::orch_locks::OrchTryLock,
 };
 use foundation::prelude::*;
