+++ src/orchestration/src/prelude.rs
@@ -12,8 +12,8 @@
 //
 
 pub use crate::actions::action::*;
+pub use crate::actions::catch::*;
 pub use crate::actions::concurrency::*;
-pub use crate::actions::event::*;
 pub use crate::actions::invoke::*;
 pub use crate::actions::sequence::*;
 pub use crate::actions::sync::*;
